codemirror fixes
